requirements.txt
fastapi==0.116.1
uvicorn==0.35.0

# Database
SQLAlchemy==2.0.43
psycopg2-binary==2.9.10

# Auth & Security
python-jose==3.5.0
passlib[bcrypt]==1.7.4

# Validation & environment
pydantic==2.11.7
python-dotenv==1.1.1

# Request parsing
python-multipart==0.0.20

# Optional but useful
email-validator==2.3.0
